+++ packages/elements/table-tree/examples/render-prop-async.js
@@ -25,7 +25,7 @@ export default () => (
     <Rows
       items={getItemsData}
       render={({ title, numbering, page, children }) => (
-        <Row key={numbering} hasChildren={children.length > 0}>
+        <Row itemId={numbering} hasChildren={children.length > 0}>
           <Cell singleLine>{title}</Cell>
           <Cell singleLine>{numbering}</Cell>
           <Cell singleLine>{page}</Cell>