+++ packages/elements/table-tree/src/components/Headers.js
@@ -1,7 +1,7 @@
 // @flow
 import React, { PureComponent, type Element, type ChildrenArray } from 'react';
 import Header from './Header';
-import { TreeRowContainer } from '../styled';
+import { HeadersContainer } from '../styled';
 
 type Props = {
   children: ChildrenArray<Element<typeof Header>>,
@@ -10,11 +10,11 @@ type Props = {
 export default class Headers extends PureComponent<Props> {
   render() {
     return (
-      <TreeRowContainer>
+      <HeadersContainer role={'row'}>
         {React.Children.map(this.props.children, (header, index) =>
           React.cloneElement(header, { key: index, columnIndex: index }),
         )}
-      </TreeRowContainer>
+      </HeadersContainer>
     );
   }
 }