+++ packages/fabric/editor-common/src/schema/nodes/doc.ts
@@ -74,12 +74,10 @@ export type TableCellContent = Array<
   | Heading
   | CodeBlock
   | MediaGroup
-  | MediaSingle
   | ApplicationCard
   | DecisionList
   | TaskList
   | Extension
-  | InlineExtension
   | BodiedExtension
 >;