+++ packages/elements/page-header/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/page-header
 
+## 2.2.0
+- [minor] Update readme and docs [7a53047](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/7a53047)
+
 ## 2.1.0
 - [minor] Add React 16 support. [12ea6e4](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/12ea6e4)