+++ packages/elements/tag/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/tag
 
+## 3.1.1
+
 ## 3.1.0
 - [minor] Add React 16 support. [12ea6e4](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/12ea6e4)