+++ packages/fabric/editor-core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @atlaskit/editor-core
 
+## 61.4.11
+
+- [patch] Insert media group instead of single image inside table [5b4aaa0](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/5b4aaa0)
+
+## 61.4.10
+
+## 61.4.9
+- [patch] Backtick should be removed at paste if its followed by code. [d74188d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d74188d)
+
+## 61.4.8
+
+
+- [patch] Add timestamp with filename on paste [18b1108](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/18b1108)
+
 ## 61.4.7